+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Dev
 
+### Bug fixes
+
+- Handle Elixir 1.12 bug on single atom ASTs
+
 ## v0.1.0 (2021-09-19)
 
 - Initial release